requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
greenlet==3.3.2
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
redis==7.3.0
requests==2.32.5
six==1.17.0
spotipy==2.26.0
SQLAlchemy==2.0.48
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
